Fix panics on unvalidated input; fuzz every input-facing API
The robustness fuzz harness only covered the Unicode algorithms. Extended it to
throw 40k random inputs (malformed locale tags, out-of-range DateTime fields,
NaN/Inf/huge f64, full-u8 time fields, extreme years) at the bidi/IDNA/spoof
analyzers and every locale-aware formatter/parser. It caught four real panics —
all the "arithmetic on uncontrolled data" class (the kind behind the Nov-2025
Cloudflare outage):

- datetime::field: month-name index `[m-1]` underflowed at month 0 and ran out
  of bounds past 12 (DateTime fields are documented as unvalidated). Clamped.
- datetime::weekday: Sakamoto's i32 arithmetic overflowed for extreme years and
  `year-1` underflowed at i32::MIN. Now computed in i64 with rem_euclid.
- datetime::field '12-hour': `hour + 11` overflowed u8 for hour > 244. Widened.
- DateTime::add_seconds: jdn*86400 + delta overflowed i64. Now saturating.
- number::format_with: `int_str.len() as u8` truncated for >255-digit values,
  letting `min_int - len` underflow. Compared in usize.

No behavior change for valid input (all existing tests pass). The two fuzz tests
run in CI under --all-features, so this class of bug can't silently return.

+++ src/number.rs
@@ -402,9 +402,11 @@ fn format_with(p: &Pattern, value: f64, decimal: &str, group: &str, minus: &str)
         None => (formatted.as_str(), ""),
     };
 
-    // Left-pad the integer to the minimum digit count.
+    // Left-pad the integer to the minimum digit count. Compare in `usize` (not
+    // `as u8`, which would truncate for >255-digit values and could underflow
+    // the subtraction below).
     let mut int_owned;
-    let int_str: &str = if (int_str.len() as u8) < p.min_int {
+    let int_str: &str = if int_str.len() < p.min_int as usize {
         int_owned = String::new();
         for _ in 0..(p.min_int as usize - int_str.len()) {
             int_owned.push('0');
